Refactored player hurt handling and it now actually ends processing once one item has taken care of the damage. Related to #300
