+++ src/main/java/xreliquary/items/ItemAngelicFeather.java
@@ -6,12 +6,18 @@ import net.minecraft.init.MobEffects;
 import net.minecraft.item.EnumRarity;
 import net.minecraft.item.ItemStack;
 import net.minecraft.potion.PotionEffect;
+import net.minecraft.util.DamageSource;
 import net.minecraft.world.World;
+import net.minecraftforge.event.entity.living.LivingAttackEvent;
 import net.minecraftforge.fml.relauncher.Side;
 import net.minecraftforge.fml.relauncher.SideOnly;
 import xreliquary.Reliquary;
+import xreliquary.handler.CommonEventHandler;
+import xreliquary.handler.IPlayerHurtHandler;
+import xreliquary.init.ModItems;
 import xreliquary.reference.Names;
 import xreliquary.reference.Settings;
+import xreliquary.util.InventoryHelper;
 
 import javax.annotation.Nonnull;
 
@@ -22,6 +28,28 @@ public class ItemAngelicFeather extends ItemBase {
 		this.setCreativeTab(Reliquary.CREATIVE_TAB);
 		this.setMaxStackSize(1);
 		canRepair = false;
+
+		CommonEventHandler.registerPlayerHurtHandler(new IPlayerHurtHandler() {
+			@Override
+			public boolean canApply(EntityPlayer player, LivingAttackEvent event) {
+				return event.getSource() == DamageSource.FALL
+						&& player.getFoodStats().getFoodLevel() > 0
+						&& InventoryHelper.playerHasItem(player, ModItems.angelicFeather)
+						&& player.fallDistance > 0.0F;
+			}
+
+			@Override
+			public boolean apply(EntityPlayer player, LivingAttackEvent event) {
+				float hungerDamage = event.getAmount() * ((float) Settings.Items.AngelicFeather.hungerCostPercent / 100F);
+				player.addExhaustion(hungerDamage);
+				return true;
+			}
+
+			@Override
+			public Priority getPriority() {
+				return Priority.HIGH;
+			}
+		});
 	}
 
 	// so it can be extended by phoenix down
