+++ src/main/java/xreliquary/items/ItemKrakenShell.java
@@ -5,11 +5,18 @@ import net.minecraft.entity.player.EntityPlayer;
 import net.minecraft.init.MobEffects;
 import net.minecraft.item.ItemStack;
 import net.minecraft.potion.PotionEffect;
+import net.minecraft.util.DamageSource;
 import net.minecraft.world.World;
+import net.minecraftforge.event.entity.living.LivingAttackEvent;
 import net.minecraftforge.fml.relauncher.Side;
 import net.minecraftforge.fml.relauncher.SideOnly;
 import xreliquary.Reliquary;
+import xreliquary.handler.CommonEventHandler;
+import xreliquary.handler.IPlayerHurtHandler;
+import xreliquary.init.ModItems;
 import xreliquary.reference.Names;
+import xreliquary.reference.Settings;
+import xreliquary.util.InventoryHelper;
 
 public class ItemKrakenShell extends ItemBase {
 
@@ -19,6 +26,27 @@ public class ItemKrakenShell extends ItemBase {
 		this.setMaxDamage(0);
 		this.setMaxStackSize(1);
 		canRepair = false;
+
+		CommonEventHandler.registerPlayerHurtHandler(new IPlayerHurtHandler() {
+			@Override
+			public boolean canApply(EntityPlayer player, LivingAttackEvent event) {
+				return event.getSource() == DamageSource.DROWN
+						&& player.getFoodStats().getFoodLevel() > 0
+						&& InventoryHelper.playerHasItem(player, ModItems.krakenShell);
+			}
+
+			@Override
+			public boolean apply(EntityPlayer player, LivingAttackEvent event) {
+				float hungerDamage = event.getAmount() * ((float) Settings.Items.KrakenShell.hungerCostPercent / 100F);
+				player.addExhaustion(hungerDamage);
+				return true;
+			}
+
+			@Override
+			public Priority getPriority() {
+				return Priority.HIGH;
+			}
+		});
 	}
 
 	@Override
